refactor(mocks): re-organize mock files
