+++ lib/core/use_cases/extras/get_tv_show_genres.dart
@@ -1,4 +1,4 @@
-import 'package:screen_pal/core/entities/genre.dart';
+import 'package:screen_pal/core/entities/extras/genre.dart';
 import 'package:screen_pal/core/repo/extras_repo.dart';
 
 class GetTvShowGenres {
