+++ test/core/use_cases/extras/get_tv_show_genres_test.dart
@@ -4,7 +4,7 @@ import 'package:mockito/mockito.dart';
 import 'package:screen_pal/core/use_cases/extras/get_tv_show_genres.dart';
 
 import '../../../helpers/dummy/dummy_genres.dart';
-import '../../../helpers/mocks/repo/repositories.mocks.dart';
+import '../../../helpers/mocks/repositories.mocks.dart';
 
 void main() {
   late GetTvShowGenres subject;
