+++ test/core/use_cases/movies/get_now_playing_movies_test.dart
@@ -3,7 +3,7 @@ import 'package:mockito/mockito.dart';
 
 import 'package:screen_pal/core/use_cases/movies/get_now_playing_movies.dart';
 
-import '../../../helpers/mocks/repo/repositories.mocks.dart';
+import '../../../helpers/mocks/repositories.mocks.dart';
 
 void main() {
   late GetNowPlayingMovies testUsecase;
